Remove live_data
