+++ packages/api/src/utils/sst-time-series.ts
@@ -2,7 +2,8 @@ import { Logger } from '@nestjs/common';
 import Bluebird from 'bluebird';
 import { In, Repository } from 'typeorm';
 import { Point } from 'geojson';
-import { flatten, groupBy, isNil, omit, times } from 'lodash';
+import { flatten, groupBy, isNil, omit, sortBy, times } from 'lodash';
+import { HistoricalMonthlyMean } from 'sites/historical-monthly-mean.entity';
 import { DateTime } from '../luxon-extensions';
 import { Site } from '../sites/sites.entity';
 import { Sources } from '../sites/sources.entity';
@@ -15,7 +16,6 @@ import {
   refreshMaterializedView,
 } from './time-series.utils';
 import { calculateAlertLevel } from './bleachingAlert';
-import { getSstAnomaly } from './liveData';
 import { ValueWithTimestamp } from './sofar.types';
 import { Metric } from '../time-series/metrics.enum';
 
@@ -29,6 +29,53 @@ interface Repositories {
 
 // Initialize Nest logger
 const logger = new Logger('SSTTimeSeries');
+
+/**
+ * The daily global 5km SSTA product requires a daily climatology to calculate the daily SST anomalies.
+ * Daily climatologies (DC) are derived from the monthly mean (MM) climatology via linear interpolation.
+ * To achieve this, we assigned the MM value to the 15th day of each corresponding month, with the individual
+ * days between these dates being derived using linear interpolation. We then calculate the SSTA product using:
+ *
+ * ST_anomaly = SST - DC
+ *
+ * where the SST is the value for the day in question, and DC is the corresponding daily climatology for that
+ * day of the year.
+ * */
+const getSstAnomaly = (
+  historicalMonthlyMean: HistoricalMonthlyMean[],
+  satelliteTemperature?: ValueWithTimestamp,
+) => {
+  if (historicalMonthlyMean.length < 12 || !satelliteTemperature?.value) {
+    return undefined;
+  }
+
+  const orderedMontlyMax = sortBy(historicalMonthlyMean, 'month');
+  const now = DateTime.now().startOf('day');
+
+  // The date of the previous value. Subtract 15 days from the current date
+  // and see in which month the result falls. The date we are looking for is
+  // the 15th day of this month.
+  const previousDate = now.minus({ days: 15 }).set({ day: 15 }).startOf('day');
+
+  // The date of the next value. It must fall on the next month of the previous
+  // value.
+  const nextDate = previousDate.plus({ months: 1 });
+
+  // We can index `orderedMontlyMax` with `DateTime.get('month')` since it returns
+  // a value between 1 and 12
+  const previousValue =
+    orderedMontlyMax[previousDate.get('month') - 1].temperature;
+  const previousDistance = now.diff(previousDate, 'days').days;
+  const nextValue = orderedMontlyMax[nextDate.get('month') - 1].temperature;
+  const nextDistance = nextDate.diff(now, 'days').days;
+  const deltaDays = previousDistance + nextDistance;
+
+  const interpolated =
+    previousValue * (1 - previousDistance / deltaDays) +
+    nextValue * (1 - nextDistance / deltaDays);
+
+  return satelliteTemperature.value - interpolated;
+};
 
 /**
  * Get sites entities based on the given siteIds array.
